pyproject.toml
[project]
name = "clip-model-serving-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.120.1",
    "greenlet>=3.2.4",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "python-jose[cryptographyl]>=3.5.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "safetensors>=0.4.0",
    "sqlalchemy>=2.0.44",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "psycopg>=3.2.12",
    "pytest>=8.4.2",
    "ruff>=0.14.2",
    "sqlacodegen>=3.0.0",
    "types-requests>=2.32.4.20250913",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "D104",  # Missing docstring in public package
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow re-exports without __all__
